Changed the logic of the pagination block
Now, if the number of pages is equal to one, then the pagination block is displayed completely, but all buttons are disabled. If there are two pages, then two buttons with numbers one and two are displayed. If there are three or more pages, then the pagination block works by default. Also moved the logic of the work of individual elements into separate functions. And edited the styles to correct display if the number of pages  is equal to two